ref(Logging): move native logging callbacks into `Util/Logging`
