+++ Assets/Gothic-UnZENity-Core/Scripts/Util/Logging.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 8b832fa42bbf457098721978a1b8a232
+timeCreated: 1718731626
